requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langsmith==0.4.13
neo4j==5.28.2
openai==1.99.9
orjson==3.11.1
packaging==25.0
playwright==1.54.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
pytz==2025.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
tree-sitter==0.25.1
tree-sitter-python==0.23.6
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zstandard==0.23.0
chromadb
langchain-community
